API LGTI/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
greenlet==3.0.1
idna==3.4
packaging==23.2
python-dotenv==1.0.0
pyzabbix==1.3.0
requests==2.31.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
whitenoise==6.6.0